data-fabric-platform/requirements.txt
# Core Data Processing
pyspark==3.5.0
delta-spark==3.0.0
pyarrow==14.0.1
pandas==2.1.4
numpy==1.26.2

# Cloud Providers - Azure
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-mgmt-datafactory==4.0.0
azure-mgmt-synapse==2.1.0
azure-monitor-query==1.2.0

# Cloud Providers - AWS
boto3==1.34.17
awswrangler==3.5.2
s3fs==2024.2.0

# Hadoop & Big Data
hdfs==2.7.3
pyhive==0.7.0
thrift==0.16.0
pyhdfs==0.3.1

# Data Quality & Validation
great-expectations==0.18.8
pandera==0.17.2
cerberus==1.3.5

# Orchestration
apache-airflow==2.8.1
apache-airflow-providers-apache-spark==4.6.0
apache-airflow-providers-microsoft-azure==8.4.0
apache-airflow-providers-amazon==8.16.0

# Database Connectors
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymongo==4.6.1
pymysql==1.1.0

# Monitoring & Logging
prometheus-client==0.19.0
statsd==4.0.1
python-json-logger==2.0.7

# API & Web
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0

# Data Serialization
avro-python3==1.10.2
fastavro==1.9.3
protobuf==4.25.2

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
moto==5.0.0  # AWS mocking
responses==0.24.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
python-dateutil==2.8.2
jinja2==3.1.3

# Development Tools
black==24.1.1
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0
pre-commit==3.6.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Infrastructure as Code
python-terraform==0.10.1

# Data Lineage & Catalog
apache-atlas==0.0.14
amundsen-databuilder==7.0.0
